Move StackFrame into Function
